+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - General:
   - Same instance is reused if possible
+  - Added refresh button in main toolbar (or use F5)
 - Bug fixes:
   - Fixed unwanted empty lines in combos
   - Fixed text selection in search results